test: a user can not delete himself.

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -113,6 +114,18 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         //
+
+        //evitar que el usuario logueado se borre a si mismo
+        if($user->id == Auth::user()->id){
+            session()->flash('swal',[
+                'icon' => 'error',
+                'title' => 'No puedes eliminarte a ti mismo',
+                'text' => 'No puedes eliminar tu propio usuario',
+            ]);
+            abort(403, 'No puedes eliminar tu propio usuario');
+            return redirect(route('admin.users.index'));   
+        }
+
         //elimina los roles asociados al usuario antes de eliminarlo
         $user->roles()->detach();
         //elimina al usurio de la bd
